+++ examples/antd-calendar.js
@@ -1,6 +1,7 @@
 import 'rc-calendar/assets/index.less';
 import React from 'react';
-import Calendar,{Picker as DatePicker} from 'rc-calendar';
+import Calendar from 'rc-calendar';
+import DatePicker from 'rc-calendar/src/Picker';
 import zhCn from 'gregorian-calendar/lib/locale/zh-cn'; // spm error
 import DateTimeFormat from 'gregorian-calendar-format';
 import GregorianCalendar from 'gregorian-calendar';
@@ -8,10 +9,11 @@ import CalendarLocale from 'rc-calendar/src/locale/zh-cn';
 var now = new GregorianCalendar(zhCn);
 now.setTime(Date.now());
 
+const formatter = new DateTimeFormat('yyyy-MM-dd HH:mm:ss');
+
 var defaultCalendarValue = new GregorianCalendar(zhCn);
 defaultCalendarValue.setTime(Date.now());
 defaultCalendarValue.addMonth(-1);
-
 
 function disabledDate(current, value) {
   var date = new Date();
@@ -22,12 +24,12 @@ function disabledDate(current, value) {
 }
 
 var Test = React.createClass({
-  handleChange(value) {
-    console.log('DatePicker change: ' + (value && this.props.formatter.format(value)));
+  onChange(value) {
+    console.log('DatePicker change: ' + (value && formatter.format(value)));
   },
 
-  handleCalendarSelect(value) {
-    console.log('calendar select: ' + (value && this.props.formatter.format(value)));
+  onCalendarSelect(value) {
+    console.log('calendar select: ' + (value && formatter.format(value)));
     // controlled value
     this.setState({
       time: Date.now(),
@@ -35,19 +37,13 @@ var Test = React.createClass({
     });
   },
 
-  handleCalendarOk(value) {
-    console.log('calendar ok: ' + (value && this.props.formatter.format(value)));
+  onCalendarOk(value) {
+    console.log('calendar ok: ' + (value && formatter.format(value)));
     // controlled value
     this.setState({
       time: Date.now(),
       value: value
     });
-  },
-
-  getDefaultProps() {
-    return {
-      formatter: new DateTimeFormat('yyyy-MM-dd HH:mm:ss')
-    }
   },
 
   getInitialState() {
@@ -59,7 +55,7 @@ var Test = React.createClass({
     };
   },
 
-  handleShowTimeChange(e) {
+  onShowTimeChange(e) {
     this.setState({
       showTime: e.target.checked
     });
@@ -80,13 +76,13 @@ var Test = React.createClass({
                              showTime={this.state.showTime}
                              showOk={true}
                              disabledDate={disabledDate}
-                             onOk={this.handleCalendarOk}
-                             onSelect={this.handleCalendarSelect}
-                             onClear={this.handleCalendarSelect.bind(this, null)} showClear={true}/>;
+                             onOk={this.onCalendarOk}
+                             onSelect={this.onCalendarSelect}
+                             onClear={this.onCalendarSelect.bind(this, null)} showClear={true}/>;
     return <div style={{width: 240, margin: 20}} data-time={this.state.time}>
       <div style={{marginBottom:10}}>
         <span>
-          <input type='checkbox' checked={this.state.showTime} onChange={this.handleShowTimeChange}/>
+          <input type='checkbox' checked={this.state.showTime} onChange={this.onShowTimeChange}/>
           showTime
         </span>
         &nbsp;&nbsp;&nbsp;&nbsp;
@@ -101,23 +97,56 @@ var Test = React.createClass({
       }}>
         <DatePicker
           adjustOrientOnCalendarOverflow={true}
+          adjustOrientOnCalendarOverflow={true}
           animation="slide-up"
           disabled={state.disabled}
-          trigger={<span className="rc-calendar-picker-icon" />}
-          formatter={this.props.formatter} calendar={calendar}
-          value={state.value} onChange={this.handleChange}>
-          <input className="rc-calendar-picker-input" style={{width:200}} disabled={state.disabled}
-                 placeholder="请选择日期"/>
+          calendar={calendar}
+          value={state.value}
+          onChange={this.onChange}>
+          {
+            ({value})=> {
+              return (
+                <span>
+                <input placeholder="请选择日期" style={{width:250}}
+                       disabled={state.disabled}
+                       className="ant-calendar-picker-input ant-input"
+                       value={value && formatter.format(value)}/>
+                <span className="ant-calendar-picker-icon" unselectable="true"/>
+                </span>
+              );
+            }
+          }
         </DatePicker>
       </div>
     </div>;
   }
 });
 
+function onStandaloneSelect(value) {
+  console.log('onStandaloneSelect');
+  console.log(formatter.format(value))
+}
+
+function onStandaloneChange(value) {
+  console.log('onStandaloneChange');
+  console.log(formatter.format(value))
+}
+
+
 React.render(<div style={{zIndex:1000,position:'relative'}}>
+  <link href="http://ant.design/dist/antd.css" rel="stylesheet" type="text/css"/>
   <h2>zh-cn</h2>
 
-  <div style={{marginTop:200,width:600}}>
+  <div style={{width:600}}>
+    <div style={{margin:10}}>
+      <Calendar showWeekNumber={false}
+                locale={CalendarLocale}
+                defaultValue={now}
+                onChange={onStandaloneChange}
+                disabledDate={disabledDate}
+                onSelect={onStandaloneSelect}
+                showTime={true}/>
+    </div>
     <div style={{float:'left',width:300}}>
       <Test defaultValue={now}/>
     </div>